refactor(transfer): rename transfer to headless
