+++ __tests__/transfer/store/transfer.spec.ts
@@ -1,7 +1,7 @@
 import { assert } from "chai";
 import { GraphQLClient } from "graphql-request";
 import { getSdk } from "src/generated/graphql-request";
-import TransferStore from "src/transfer/stores/transfer";
+import HeadlessStore from "src/transfer/stores/headless";
 
 const client = new GraphQLClient(`http://localhost:23061/graphql`);
 const headlessGraphQLSDK = getSdk(client);
@@ -11,7 +11,7 @@ describe("snapshot", function () {
 
   describe("get current epoch from store", function () {
     it("should be equal base epoch with empty store", async function () {
-        const transferStore = new TransferStore(headlessGraphQLSDK);
+        const transferStore = new HeadlessStore(headlessGraphQLSDK);
         await transferStore.trySetAgentAddress();
         console.log(transferStore.getAgentAddress());
     });
