+++ src/transfer/App.tsx
@@ -6,22 +6,22 @@ import { getSdk } from "src/generated/graphql-request";
 import { IStoreContainer } from "src/interfaces/store";
 import { ITransferStoreContainer, StoreContext } from "./hooks";
 import MainPage from "./pages/main/main";
-import MenuStore from "./stores/menu";
-import TransferStore from "./stores/transfer";
+import MenuStore from "./stores/views/menu";
+import HeadlessStore from "./stores/headless";
 
 const client = new GraphQLClient(`http://localhost:23061/graphql`);
 const headlessGraphQLSDK = getSdk(client);
 
 const storeContainer: ITransferStoreContainer = {
-  transferStore: new TransferStore(headlessGraphQLSDK),
+  headlessStore: new HeadlessStore(headlessGraphQLSDK),
   menuStore: new MenuStore()
 }
 
 const App: React.FC = () => {
   useEffect(() => {
     async function main() {
-      await storeContainer.transferStore.trySetAgentAddress();
-      await storeContainer.transferStore.updateBalance();
+      await storeContainer.headlessStore.trySetAgentAddress();
+      await storeContainer.headlessStore.updateBalance();
     }
     main();
   }, []);
