+++ src/transfer/pages/exchange/exchange.tsx
@@ -1,6 +1,6 @@
 import { T } from "@transifex/react";
 import React from "react"
-import { MenuItems } from "src/transfer/stores/menu";
+import { MenuItems } from "src/transfer/stores/views/menu";
 import useStores from "../../../transfer/hooks";
 
 const transifexTags = "Transfer/Exchange";
