+++ src/transfer/pages/transfer/transfer.tsx
@@ -1,35 +1,43 @@
 import { T } from "@transifex/react";
+import { shell } from "electron";
 import { observer } from "mobx-react";
 import React, { useContext } from "react"
 import { useState } from "react";
+import SendingDialog from "src/transfer/components/SendingDialog/SendingDialog";
 import { StoreContext } from "src/transfer/hooks";
 
 const transifexTags = "Transfer/Transfer";
 
 const TransferPage: React.FC = observer(() => {
-  const { transferStore } = useContext(StoreContext);
+  const { headlessStore } = useContext(StoreContext);
   const [recipient, setRecipient] = useState<string>('');
-  const [amount, setAmount] = useState<number>(0);
   const [memo, setMemo] = useState<string>('');
+  const [tx, setTx] = useState<string>('');
+  const [amount, setAmount] = useState<number>(0);
   const [isSending, setIsSending] = useState<boolean>(false);
 
   const handleButton = async (event: React.MouseEvent<HTMLButtonElement>) => {
     event.preventDefault();
     setIsSending(true);
-    const tx = await transferStore.transferGold(recipient, amount, memo);
+    const tx = await headlessStore.transferGold(recipient, amount, memo);
+    setTx(tx);
 
-    transferStore.confirmTransaction(tx, undefined, (blockIndex, blockHash) => {
-      console.log(`Block #${blockIndex} (${blockHash})`);
-      setIsSending(false);
-    },
-    (blockIndex, blockHash) => {
-      console.log(`Failed`);
-      setIsSending(false);
-    },
-    (blockIndex, blockHash) => {
-      console.log(`Timeout`);
-      setIsSending(false);
-    });
+    headlessStore.confirmTransaction(tx, undefined,
+      (blockIndex, blockHash) => {
+        console.log(`Block #${blockIndex} (${blockHash})`);
+        setIsSending(false);
+        setTx('');
+      },
+      (blockIndex, blockHash) => {
+        console.log(`Failed`);
+        setIsSending(false);
+        setTx('');
+      },
+      (blockIndex, blockHash) => {
+        console.log(`Timeout`);
+        setIsSending(false);
+        setTx('');
+      });
   }
 
   return (
@@ -54,9 +62,9 @@ const TransferPage: React.FC = observer(() => {
           <T _str="Enter the amount of NCG to send." _tags={transifexTags} />
         </p>
         <p>
-          <T _str="(Your balance: {ncg} NCG)" _tags={transifexTags} ncg={transferStore.balance}/>
+          <T _str="(Your balance: {ncg} NCG)" _tags={transifexTags} ncg={headlessStore.balance} />
         </p>
-        <button onClick={() => transferStore.updateBalance()}>refresh</button>
+        <button onClick={() => headlessStore.updateBalance()}>refresh</button>
         <input onChange={e => setAmount(parseFloat(e.target.value))}></input>
       </div>
       <div>
@@ -71,6 +79,15 @@ const TransferPage: React.FC = observer(() => {
       <div>
         <button onClick={handleButton} disabled={isSending}>Send</button>
       </div>
+      <SendingDialog
+        open={tx !== ''}
+        onDetailedView={() => {
+          shell.openExternal(
+            `https://explorer.libplanet.io/9c-main/transaction/?${tx}`
+          );
+        }}
+      />
+
     </div>
   );
 });
